% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_filter_allSpec.R
\name{re_filter_allSpec}
\alias{re_filter_allSpec}
\title{Eliminate restriction enzyme produced non-unique DNA fragmentation to all species}
\usage{
re_filter_allSpec(SimScore)
}
\arguments{
\item{SimScore}{-data.frame, containing score value from re_sim}
}
\value{
data.frame containing list of restriction enzyme which are unique to all species
}
\description{
Eliminate restriction enzyme produced non-unique DNA fragmentation to all species
}
\examples{
Homfasta<-read.fasta("P.homarus.fasta")
SimHom<-re_sim(fasta = Homfasta, min.bp = 150)

Ornfasta<-read.fasta("P.ornatus.fasta")
SimOrn<-re_sim(fasta = Ornfasta, min.bp = 150)

Polfasta<-read.fasta("P.polyphagus.fasta")
SimPol<-re_sim(fasta = Polfasta, min.bp = 150)

MergeScore<- data.frame(SimHom$Score, SimOrn$Score, SimPol$Score)
rownames(MergeScore)<-rownames(SimHom)

dat<- re_filter_allSpec(SimScore = MergeScore)
}
