% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_filter_fragnumb.R
\name{re_filter_fragnumb}
\alias{re_filter_fragnumb}
\title{Keep restriction enzymes produced desired fragment number only}
\usage{
re_filter_fragnumb(EnzList, SimTarget, FragNumb)
}
\arguments{
\item{EnzList}{-data.frame, list of enzyme after re_filter_target}

\item{SimTarget}{-data.frame, list of initial enzyme}

\item{FragNumb}{-numeric, maximum fragment number allowed}
}
\value{
data.frame containing list of restriction enzyme produced desired fragment number
}
\description{
Keep restriction enzymes produced desired fragment number only
}
\examples{
?re_filter_targetSpec
Filt3.2Frag<-re_filter_fragnumb(EnzList = Filt2.Target, SimTarget = SimHom, FragNumb = 2)
}
