% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_filter_targetSpec.R
\name{re_filter_targetSpec}
\alias{re_filter_targetSpec}
\title{Eliminate restriction enzyme produced non-unique DNA fragmentation to the target species}
\usage{
re_filter_targetSpec(x)
}
\arguments{
\item{x}{-data.frame from re_filter_lowfrag or score data from re_sim}
}
\value{
data.frame containing list of restriction enzyme produced unique DNA fragmentation  to the target species
}
\description{
Eliminate restriction enzyme produced non-unique DNA fragmentation to the target species
}
\examples{
?re_filter_lowfrag
Filt2.Target<-re_filter_target(Filt1.DataMerge)
}
