% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_pcr.R
\name{re_pcr}
\alias{re_pcr}
\title{Estimate the PCR product}
\usage{
re_pcr(x, Forward, Reverse, id, out.fasta)
}
\arguments{
\item{x}{-DNAstring data from re_prep, complete genome}

\item{Forward}{-character, forward primer 5' direction}

\item{Reverse}{-character, Reverse primer 5' direction}

\item{id}{-character, species identity}

\item{out.fasta}{-character, file name of the output fasta file}
}
\value{
print of the DNAstring and fasta file output
}
\description{
Estimate the PCR product
}
\examples{
Fwd<- "GGTCAACAAATCATAAAGATATTG"
Rev<-"TAAACTTCAGGGTGACCAAAAAATCA"
mtDNAHom<-re_prep("NC_016015.1", fasta.out = "mtDNAHom.fasta")
HomAmpl<-re_pcr(mtDNAHom, Fwd,Rev, "P.homarus","P.homarus.fasta")
}
