% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_prep.R
\name{re_prep}
\alias{re_prep}
\title{Fetch DNA sequences from NCBI and prepare the DNAstring data}
\usage{
re_prep(x, fasta.out)
}
\arguments{
\item{x}{-character, NCBI accesion number of the targeted sequence or NCBI advanced search keyword}

\item{fasta.out}{-character, file name of the output fasta file}
}
\value{
print of the DNAstring and fasta file output
}
\description{
Fetch DNA sequences from NCBI and prepare the DNAstring data
}
\examples{
mtDNAHom<-re_prep("NC_016015.1", fasta.out = "mtDNAHom.fasta")
}
