% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_sim.R
\name{re_sim}
\alias{re_sim}
\title{Simulate Restriction Enzyme digestion}
\usage{
re_sim(fasta, min.bp)
}
\arguments{
\item{fasta}{-DNAstring data, from re_pcr, PCR product or partial sequence}

\item{min.bp}{-numeric, minimum length of produced DNA fragment after enzyme digestion}
}
\value{
list containing digestion score and fragment number produced from enzyme digestion
}
\description{
Simulate Restriction Enzyme digestion
}
\examples{
#Homfasta<-read.fasta("P.homarus.fasta")
SimHom<-re_sim(fasta = Homfasta, min.bp = 150)
}
